move types into separate files and namespaces
