+++ DVDispatcherMod/DispatcherHintShowers/IDispatcherHintShower.cs
@@ -1,0 +1,7 @@
+﻿using DVDispatcherMod.DispatcherHints;
+
+namespace DVDispatcherMod.DispatcherHintShowers {
+    public interface IDispatcherHintShower {
+        void SetDispatcherHint(DispatcherHint dispatcherHintOrNull);
+    }
+}
