+++ DVDispatcherMod/HarmonyPatches/Constants.cs
@@ -1,0 +1,6 @@
+﻿namespace DVDispatcherMod.HarmonyPatches {
+    public static class Constants {
+        public const float START_WIDTH = 0.0025f; // Original = 0.005f;
+        public const float END_WIDTH = 0.01f; // Original = 0.02f;
+    }
+}
